services/performance-optimizer.ts
import OpenAI from 'openai';
import { createClient } from '@/utils/supabase/client';

interface PerformanceMetrics {
  responseTime: number;
  successRate: number;
  errorCount: number;
}

interface PerformanceAnalysis {
  metrics: PerformanceMetrics;
  recommendations: string[];
}

export class PerformanceOptimizer {
  public static supabase = createClient();

  /**
   * Track performance metrics
   */
  public static async trackMetrics(metrics: PerformanceMetrics): Promise<void> {
    const { error } = await this.supabase.from('performance_metrics').insert({
      response_time: metrics.responseTime,
      success_rate: metrics.successRate,
      error_count: metrics.errorCount,
      timestamp: new Date().toISOString()
    });

    if (error) {
      throw new Error(`Failed to store metrics: ${error.message}`);
    }
  }

  /**
   * Analyze performance data and provide recommendations
   */
  public static async analyzePerformance(): Promise<PerformanceAnalysis> {
    const { data, error } = await this.supabase
      .from('performance_metrics')
      .select('*')
      .order('timestamp', { ascending: false })
      .limit(100);

    if (error) {
      throw new Error(`Failed to fetch metrics: ${error.message}`);
    }

    // Calculate aggregate metrics
    const metrics = this.calculateAggregateMetrics(data);

    // Get recommendations
    const recommendations = await this.generateRecommendations(metrics);

    return {
      metrics,
      recommendations
    };
  }

  private static calculateAggregateMetrics(data: any[]): PerformanceMetrics {
    return {
      responseTime:
        data.reduce((acc, m) => acc + m.response_time, 0) / data.length,
      successRate:
        data.reduce((acc, m) => acc + m.success_rate, 0) / data.length,
      errorCount: data.reduce((acc, m) => acc + m.error_count, 0)
    };
  }

  private static async generateRecommendations(
    metrics: PerformanceMetrics
  ): Promise<string[]> {
    // Implementation here
    return ['Optimize database queries', 'Add caching layer'];
  }
}
